move common duplicate support to shared module outside apps (#5278)
